search menu implementation and VideoDetails

+++ app/src/main/java/com/omt/omtest/domain/Availabilities.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class Availabilities (
         @SerializedName("images") val images : List<Images>,
         @SerializedName("videoId") val videoId : String,
-        @SerializedName("startTime") val startTime : Int,
-        @SerializedName("endTime") val endTime : Int,
+        @SerializedName("startTime") val startTime : Long,
+        @SerializedName("endTime") val endTime : Long,
         @SerializedName("categories") val categories : List<Categories>,
         @SerializedName("serviceId") val serviceId : String
 )
